terraform: autoscale k8s deployments
To spare ourselves the hassle of manually resizing deployments, we
configure Kubernetes to automatically resize `intake-batch` and
`aggregate` worker pools. This requires deploying the Stackdriver custom
metrics adapter to make PubSub metrics visible to a Kubernetes
Horizontal Pod Autoscaler, then configuring an HPA for each deployment
that consults the PubSub `num_undelivered_messages` metric. See
`terraform/README.md` for more details and discussion of config
parameter choices.

This commit also modifies the `integration-tester` deployment so that it
emits more batches, forcing dev and staging clusters to exercise the
autoscaling feature. We also amend the alert for `intake-batch` task
queue size: we no longer expect that queue to periodically empty since
we have configured Kubernetes to keep it at a steady state of ~150
messages.

Resolves #484