workflow-manager: label metrics by aggregation ID
Several existing metrics in `workflow-manager` are changed from `Gauge`
to `GaugeVec` in order to admit an `aggregation_id` label. This permits
us to count various metrics broken out by aggregation ID. We also add a
number of new metrics for things like counts of ingestion, own
validation and peer validation batches found (including incomplete batches
that are ignored. This commit also deletes the `monitor` module which
provided a dummy implementation of a Prometheus Gauge. Instead of using
dummy collectors, we can simply conditionalize the push of metrics on
whether or not a push gateway URL was provided to `workflow-manager`.
This does mean that `workflow-manager` must incur the memory footprint
of the metrics collectors regardless of whether or not they are used,
but:
 (1) the footprint is quite small since we are talking about a few
     dozen gauges
 (2) as far as I know, all existing `workflow-manager` deployments do
     use Prometheus and provide a pushgateway URL
 (3) this is much simpler than implementing mocks for every metric type
     we might ever use

Resolves #338