terraform: relax prometheus alerting rules
We are observing a lot of unactionable alerts for task queue growth. On
the intake side, these represent periodic but normal spikes in incoming
ingestion batches which can be handled without additional capacity but
not within 30 minutes. On the aggregate side, it just seems that
aggregations are taking a long time, especially in us-ca-apple. In
anticipation of deploying autoscaling and parallelized aggregations,
which will amelioriate both of these situations, this commit relaxes the
alerting rules to allow more time for workers to catch up.