facilitator: refactor sample generation
Refactors the `generate_ingestion_sample()` free function in `mod
sample` so that we instead have a `SampleGenerator` struct. This brings
this module in line with its peers `intake::BatchIntaker` and
`aggregation::BatchAggregator`, and more importantly, will make it
easier to associate a `slog::Logger` with a sample generation task in
the future.