facilitator: use slog for structured logging
We adopt the `slog-rs` family of crates for structured logging. `slog`
allows us to create trees of `Loggers` and `Drains`. `Logger`s can be
annotated with key-value pairs and then every message logged through
that `Logger` will themselves be annotated with the k-v pairs. Further,
the `slog` macros `info`, `error`, etc. allow adding arbitrary
additional k-v pairs to individual events. This lets us emit structured
log messages which can then be searched in something like the GCP Log
Explorer. `Drain`s can be configured to format log records in different
manners. We use `slog_json` to emit structured JSON logs, `slog_term` to
pretty print messages to a tty or a much simpler decorator for use in
tests.

`slog`'s model is that you create a root `Logger`, and then create a
tree of child loggers from it, each of which can be specialized with
context specific to some module or task. For instance, in this commit,
we augment `aggregation::BatchAggregator` so that it creates a `Logger`
from a provided parent logger, and annotates it with a trace ID and
other interesting values.

To ease the transition from the existing logging model to `slog`, we use
`slog_scope` as a drop-in replacement for the `log`/`env_logger` combo
previously employed in `facilitator`. Subsequent PRs will plumb `Logger`
instances into each module of `facilitator` and eventually `slog_scope`
should be removed altogether.

This commit is a start on #546