task-replayer: simple utility for replaying tasks
We use message queues (GCP PubSub or AWS SNS/SQS) to manage tasks in
`prio-server`. After a task is retried five times, it gets delivered to
a dead-letter queue for triage. In some cases, the thing to do is just
to replay the failed task into the queue so it can be tried again,
perhaps because a change has been deployed that will now let the task
succeed.

This commit adds a Go utility `task-replayer` which permits operators to
replay a task from a JSON file. See `task-replayer/README.md` or
`task-replayer --help` for details.

Relates to #290