facilitator: introduce `http::RetryingAgent`
In order to be more robust against transient failures in cloud APIs, we
introduce a `RetryingAgent` which uses the `retries` crate to resend
failed HTTP requests with exponential backoff if the failure is
determined to be transient (e.g., an HTTP 503). The `RetryingAgent` is
adopted throughout the crate, meaning we now get retries when:

  - fetching peer JSON manifests
  - pulling messagesfrom GCP PubSub subscriptions
  - reading or writing to Google Cloud Storage
  - obtaining Oauth tokens from the GKE metadata service or IAM API

We should also record Prometheus metrics about retryable request
failures, which will happen in a future PR.

This change doesn't help with AWS API clients (S3, SQS and IAM) because
they do not use our `ureq` based HTTP client but rather Rusoto (and that
in turn uses Hyper). A subsequent PR will refactor
`aws_credentials::retry_request` and `aws_credentials::retryable` to be
built on `retries::retry_request`.

Related to #351, #352, #353