actions: invoke cargo-clippy directly
We previously were using the
[clippy-check](https://github.com/actions-rs/clippy-check) GitHub Action
to run clippy on changes. There's two problems with this:

 1 - Because our `Cargo.toml` isn't at the root of the repo,
   `clippy-check` can't post clippy results into the PR as comments
   (see #69 and
   https://github.com/actions-rs/clippy-check/issues/28#issuecomment-680401392)
 2 - We recently set `GITHUB_TOKEN` to be readonly at the
   `abetterinternet` level to mitigate potential security problems.
   `clippy-check` requires write permissions to post results.

I'd prefer to keep the `GITHUB_TOKEN` passed to actions readonly, and
since the results wrangling never worked anyway, we can opt out of it.

Now, instead of using the action, we just run `cargo clippy` directly
and fail the build if any warnings are found. Previously `clippy-check`
would just emit findings into the void, never to be seen by anyone.