deploy-tool: refactoring and tests
Introduces a high-level test of the `createManifests` function in
`deploy-tool/main.go` to ensure that key generation is handled correctly
in cases where we are deploying a new instance, but a packet encryption
key already exists.

To achieve this, we introduce a handful of new interfaces to mock out
interactions like uploading manifests to cloud storage or accessing
Kubernetes secrets.

Resolves #756