deploy-tool: post manifests to S3 buckets
Teaches `deploy-tool` to manage data share processor specific manifests
in AWS S3 buckets, as well as the GCS buckets it already knew how to
deal with. To support this, `deploy-tool` needs to be provided the AWS
region of the bucket and a profile to obtain credentials for AWS, so
this commit introduces just enough Terraform changes so that further
changes that will land for #655 can express those values.

Resolves #694