facilitator: use GCP SA to assume IAM role
`facilitator` was already capable of assuming AWS IAM roles when running
in GKE using `sts:AssumeRoleWithWebIdentity`, but it required that the
GCP SA mapped to the `facilitator`'s Kubernetes service account be
permitted to assume the role. This change enables `facilitator` to be
configured with an AWS IAM role to assume _and_ a GCP service account to
impersonate to do it. `facilitator` will get an _access_ token for the
default GCP service account, then use that to obtain an _identity_ token
for the specified GCP SA, then finally use that in a request to
`sts:AssumeRoleWithWebIdentity`.

On the way there, we do some cleanup in `gcp_oauth.rs` to clearly
distinguish between code that handles access tokens and code that
handles OpenID Connect identity/ID tokens, as previously all of these
were ambiguously referred to as "tokens" or "Oauth tokens".

Finally, we tweak the audience/`aud` value used in OIDC ID tokens:
previously we would use `sts.amazonaws.com/<aws-account-id>`, but this
doesn't achieve much and will make #747 more of a headache, so we now
use a fixed string instead when creating the role. Besides Terraform and
`facilitator`, this is reflected in the change to
`workflow-manager/aws/aws.go`.

Part of #747