facilitator: beef up config::Identity type
This commit enhances the existing `config::Identity` type to make it
into a struct wrapping an `Option<String>`. The goal is to abstract away
the special handling of the value `Some("")` as being equivalent to
`None` and make it easier to use `Identity` throughout the crate instead
of passing AWS role ARNs and GCP service account emails as
`Option<&str>` or `Option<String>` in various places. This will make
things more clear when a subsequent commit introduces support for AWS
authentication flows that first require authenticating to GCP, and so
`aws_credentials::Provider` will have to manage more than one identity.

We provide `std::str::FromStr` and `serde::Deserialize` implementations
on `config::Identity` to allow for graceful instantiation of `Identity`
from program arguments or JSON manifests. Along the way, we make some
improvements to the defintions in `manifest.rs` to use `Identity` and
`StoragePath` instead of `Option<String>` and `String`.