facilitator: fix GCP workload identity pool args
PR #728 added support for GCP Workload Identity Pool to federate
identity from AWS to GCP. It added a
`gcp-workload-identity-pool-provider` command line argument, but failed
to wire it up to the various `facilitator` subcommands that need it.
Besides that, we need allow Terraform to pass
`GCP_WORKLOAD_IDENTITY_POOL_PROVIDER = ""` when running `facilitator`
(it's easier to handle this in Rust than to conditionally add values to
Terraform maps), so we add a `WorkloadIdentityPoolParameters::new`
function that looks for the empty string and treats it as equivalent to
the argument not having been specified.