terraform: EKS support in account_mapping module
module `account_mapping` is responsible for creating a Kubernetes
service account and a GCP service account and then configuring both so
that the former can impersonate the latter using GKE Workload Identity
(not to be confused with GCP Workload Identity *Pool* which we use
elsewhere). This commit extends module `account_mapping` so it can also
map a Kubernetes service account to an AWS IAM role via EKS' IAM Roles
for Service Accounts feature.