terraform: introduce eks module
Adds a module that configures EKS clusters on AWS, as well as a VPC and
subnets that put all the worker nodes on private networks but with
egress to the Internet through a NAT gateway. The output variables are
harmonized with `modules/gke/gke.tf` so that `main.tf` can "generically"
get Kubernetes cluster credentials to provide to other modules. We don't
yet plumb instantiation of the EKS module in `main.tf`.