terraform: flesh out S3 cloud storage
The existing `cloud_storage_aws` module only supported creating
ingestion and peer validation buckets in S3, to exercise authenticating
to AWS from GKE as well as the S3 client. Now, we expand the module so
that it creates all of a data share processor's buckets in S3.

We also make the big IAM role switcheroo that forces the adoption of new
manifest formats: instead of the GCP DSP creating AWS IAM roles that it
will assume in order to write to the peer's S3 buckets, we now expect
the AWS DSP to create IAM roles and configure them to be assumable by a
GCP service account advertised from the GCP DSP's global manifest.