terraform: wire up server resources for devenvs
In development environments and staging, we run a pair of data share
processors configured to exchange data. We also set up a fake ingestion
server (a Kubernetes deployment that runs `facilitator
generate-sample-data-worker`) and a fake portal server (just a GCS
bucket where sum parts are emitted). These modules construct manifests
and storage for those fake servers, and this commit updates them to
handle S3 and EKS.