terraform: per-locality aggregation periods
Allow configuration of aggregation period and aggregation grace period
per locality, rather than one value for the entire env. This is needed
so that the `ta-ta` locality in the `prod-us` env can run aggregations
at a faster cadence than the rest of production.

To achieve this we use Terraform's optional object member feature, which
is currently an experiment.

Resolves #789