update prod-intl with peer DSP manifest URL
Robert over at Google let us know yesterday where the manifests for the
production instance of their data share processor would be, so this
commit updates the `prod-intl` TF vars accordingly. This change doesn't
do anything because we don't yet have any localities defined, and as
such does not need to be deployed.