facilitator: stop writing validation packets to our own storage. (#997)
* facilitator: stop writing validation packets to our own storage.

Instead, we recompute them as needed during aggregation--this works
because our validation packet is a function of the ingestion packet,
which we already read during aggregation. We still write validation
packet to our peer's storage, since knowing your peer's validation
packets is a requirement to perform validation.

* Workflow manager: schedule aggregation tasks based on intake batches.

Previously, aggregation tasks were created based on the existence of a
peer validation batch & an own validation batch. Now, they are created
based on the existence of a peer validation batch & an intake batch.
This is required since the facilitator was changed in this PR to not
write out own validation batches.

The concept of the own validation bucket still exists (it is used for
"task markers" still). It could be removed if we moved the task markers
elsewhere, but I'm leaving that for a later change.