workflow-manager: accept aggregation-window override. (#1027)
This is implemented by the new --aggregation-override-timestamp flag.
Normally (without this flag), the most recently completed aggregation
window (shifted by a grace period) will be aggregated. With this flag,
the aggregation window containing the override timestamp will be
aggregated instead.

Note that for an aggregation to actually be triggered, it still must
pass the other requirements. Specifically, the relevant validations must
have been written, and no task marker must exist for the window.