facilitator: API client metrics
Introduce `metrics::ApiClientMetricsCollector`, which measures counts of
requests to various services as well as the latency. We create a single
`ApiClientMetricsCollector` for use throughout `facilitator`, cloning it
as needed into individual API clients. At callsites, we then apply
labels for `service`, `endpoint` and `status`. `service` is meant to be
something like `s3.amazonaws.com` or `pubsub.googleapis.com`. `endpoint`
is relative to a `service`, so `GetObject` or `CreateMultipartUpload` on
`s3.amazonaws.com`, for instance. Finally, `status` is an indication of
whether the API request worked, so that we don't skew latency
measurements with a bunch of RPCs that failed due to timeouts.

`ApiClientMetricsCollector` is plumbed into `http::RetryingClient`,
which lets us capture metrics for all our GCP interactions and manifest
fetching. We separately plumb it into `aws_credentials::Provider` and
the SQS and S3 clients.

On the GCP side, we get pretty fine-grained metrics on individual HTTP
requests, since we have our own API clients written at the HTTP layer.
Things get a bit more abstract on the AWS side because we cannot peer
into the inner workings of Rusoto (at least not without forking that SDK
to introduce metrics to it).

On the AWS auth side, we coarsely measure various auth interactions
which could involve more than one HTTP transaction under the covers,
synthesizing "service" and "endpoint" names.

Finally, our metrics for storage are a bit misleading: we measure how
much time it takes to get to the initial HTTP response, but don't
measure how long it takes to read the entire body of a response. In any
case, that measurement would be misleading: we shouldn't be using a
single histogram to measure the time it takes to read response bodies
that could vary from a few KB to 100s of MB. Instead, we would want to
introduce separate metrics for something like kB/s, perhaps in
`src/batch.rs`.