workflow-manager: fix determination of aggregatable batches.
Previously, we were using the set of intake batches *over the intake
window* to determine which batches to include in an aggregation.
Instead, we should have been using the set of intake batches *over the
aggregation window* -- there is not necessarily any overlap between
these two windows!

In some cases we might be able to elide the second read of intake
batches, depending on whether there is an overlape between the intake &
aggregation windows; I leave that for future work.