Move Bucket out of storage package.
This type was overloaded: it was used as both an input to the (generic)
NewStorage function, as well as being used as a JSON deserialization
target. And it already had fields that were not associated with JSON
deserialization (i.e. KeyPrefix), filled in only by application
functionality to set some optional behavior in NewStorage.