Pipe context through storage.Manifest.
I'm not super-happy about NewManifest needing to take a context -- it
does so because the GCS API requires a context to establish a new
connection. I sorta want to make the caller provide the GCS/S3 API
client to use rather than creating one, but that won't happen in this
PR; I leave removing the Context argument from NewManifest to that later
work.