Delete manifest updater.
With https://github.com/abetterinternet/prio-server/pull/1084, I have
taken everything I want from manifest-updater. I don't think the rest
will be needed in the future, and it's a minor maintenance burden, so
`git rm` it. (It'll always be available in git history if we do need
it.)

Stacked on https://github.com/abetterinternet/prio-server/pull/1084.