prod-intl: force node group to 6 nodes
Our deploy of the four Mexican states got stuck because we ran out of
cores onto which to schedule pods, so we bump the max size of the node
group to 6. Further, we learned today that EKS doesn't automatically
deploy a cluster autoscaler for you (#1091), so we force the cluster to
six nodes.