Make key.Key an immutable type.
For now, this isn't such a big difference, but this will enable me to
add verifications/data-structure invariants later, and will enable me to
change the internal representation of the key.Key later if I want to.

Also, add a key deserialize->serialize test that matches the existing
serialize->deserialize test. Unfortuantely it's hard to be sure how
stable Go's JSON serialization will be, but I want to be very aware if
later changes to the codebase (accidentally) change the serialization
format of key.Key.