Make key.Key immutable, rename key.Raw to key.Material. (#1096)
For now, immutability isn't such a big difference, but this will enable me to add verifications/data-structure invariants later, and will enable me to change the internal representation of the key.Key later if I want to.

"Key material" is a more-standard/common name for the concept captured by key.Raw; see e.g. https://crypto.stackexchange.com/questions/80540/meaning-of-the-term-key-material.