Fix golangci-lint lints.
I'm not so happy about the `li` function I introduced -- it's to work
around the use of the LI type alias for storagetest.LocalityIngestor.
Since the type is in a different package, it doesn't like us using
implicit field initializers, since introducing a new field would break
code using implicit field initializers. But the whole point of
LocalityIngestor is to be a tuple of (locality, ingestor), so we'll
never introduce any new fields... still, I prefer `li` to trying to
disable the linter in this case.