key.Key: new API for handling versions.
Note that Key.Primary() currently selects an arbitrary primary key if
the key (incorrectly/invalidly) contains more than one primary key. This
would be problematic, but I'm about to change key.Key so that this is no
longer possible.

I had to merge the BSK/PEK tests for manifest/manifest_test.go since now
a non-empty PEK is required for the test to not panic. That's OK, too:
soon I will be adding validation that neither of these keys are empty,
so the tests needed to change before too long, anyway.