Change Key internal representation to list primary version first.
Previously, the Key APIs would permit multiple (or no) primary versions,
and it was up to the users of Key to select one to use if a single
primary was necessary. This was inherently unsafe in that different
users of the Key APIs might make different selections.

Now, we strongly require exactly one primary version for non-empty keys,
denoted in an unambiguous way, and validate that this is the case
everywhere we create new Key values. I leave the serialization format
unchanged (i.e we still include the "primary" key on the primary key
version), but this is not strictly required.