Update main_test so that it passes newly-added manifest validations.
The problem was that we were placing bogus key material into the
manifests, which caused update-keys operations to fail because we parse
manifest keys as part of pre-update validations. The tests now create
real (but deterministic) P256 keys, and the UpdateKeys validations now
effectively check that we are putting the right keys in the right place.