Add validations to key.Key type. (#1128)
* key.Key: new API for handling versions.

Note that Key.Primary() currently selects an arbitrary primary key if
the key (incorrectly/invalidly) contains more than one primary key. This
would be problematic, but I'm about to change key.Key so that this is no
longer possible.

I had to merge the BSK/PEK tests for manifest/manifest_test.go since now
a non-empty PEK is required for the test to not panic. That's OK, too:
soon I will be adding validation that neither of these keys are empty,
so the tests needed to change before too long, anyway.

* Change Key internal representation to list primary version first.

Previously, the Key APIs would permit multiple (or no) primary versions,
and it was up to the users of Key to select one to use if a single
primary was necessary. This was inherently unsafe in that different
users of the Key APIs might make different selections.

Now, we strongly require exactly one primary version for non-empty keys,
denoted in an unambiguous way, and validate that this is the case
everywhere we create new Key values. I leave the serialization format
unchanged (i.e we still include the "primary" key on the primary key
version), but this is not strictly required.

* Switch Version.CreationTime to Version.CreationTimestamp.

Previously, versions contained a CreationTime (time.Time value).
However, nearly every usage of CreationTime effectively mapped it to a
Unix timestamp first, and every test uses times that match a Unix second
timestamp exactly. But a few checks were still based on the underlying
time.Time values, and there was nothing stopping users of the key
package from creating key versions with creation times that did not
correspond to a Unix timestamp value exactly. While I'm fairly sure this
wouldn't cause problems, I'm even more sure that there are no problems
now that we are storing a Unix timestamp directly.

This will also be useful in implementing the next set of validations.

* Disallow multiple key versions with the same creation timestamp.

At a whole-program level, this effectively disallows rotating a key
multiple times per second. This is acceptable.

* After rotation, validate that key contains at least one version.

A useful invariant for later validations.

* Add RotationConfig validation ordering primary/create/delete min ages.

This (more specifically, the condition that CreateMinAge <=
DeleteMinAge) effectively stops rotations from being able to delete all
key versions -- at least one version will always live, since a rotation
operation that deleted all keys would also create a new key version by
the newly-checked-for condition, and a new key version will never be
deleted since it has age 0 and DeleteMinAge is non-negative (by existing
checks), contradicting the fact that our rotation operation deleted all
versions.

I keep the previous check as a sanity-check/backstop, but as far as I
can tell it is no longer testable.

* Add validations to manifest's UpdateKeys function (#1135)

Specifically (from #24):

* When being used to update keys in a manifest, keys must have at least one version.
* Post-update, manifests must have exactly one packet encryption key version.
* Post-update, manifests must have at least one batch signing key version.
* Post-update, manifests' non-key data must match the pre-update manifest data exactly.
* Post-update, manifests' key data for key versions that exist both pre- & post-update must match exactly.

* Add further validations to manifest's UpdateKeys function (#1136)

Specifically (from #24):

* When being used to update keys in a manifest, existing key versions' public portions should match with the public keys written to the manifest, matching by creation time/key ID (both of which would provide the same matching).
* Detection for: "if a batch signing key got so far ahead of its related manifests that the primary key was not included in the manifest, we would be trying to sign data with a key that we do not recognize as valid for verification."
* Detection for: "if a packet encryption key got so far ahead of its related manifests that the key did not include the version advertised as primary in the manifest, we would be asking clients to encrypt with a key we no longer consider valid for decryption."