Key rotator: fix bug with packet encryption key deserialization.
Previously, we weren't checking the number of bytes returned from
`Decode`; in practice, this meant packet encryption keys that
round-tripped through serialization would end up with a `D` value that
was a few bytes too long. This didn't actually break anything for the
key rotator itself, since it doesn't use the key, but as one might
imagine the facilitator was none-too-happy about receiving these
malformed keys.