Add (bogus) expiry date to manifests' batch signing public key.
This field is evidently not respected by any parties in the current ENPA
deployment (most/all of the expiration dates in the production
deployment are now in the past). But per discussion with ENPA team
members, we want to continue populating this with *something* in case
one of the parties is parsing this value & then ignoring it.