GH actions: push `key-rotator` from build step
Pushing `prio-key-rotator` container images from a step separate from
the build doesn't work because the build cache gets cleaned up between
GH action steps, and so the built image is not visible to the push
steps. We could be clever about exporting the build image, but it's
easier to just push to dockerhub from the `build` step.