Key rotator: add sanity checks to parsed keys. (#1150)
While caught in dev, the bug that led to
https://github.com/abetterinternet/prio-server/commit/8de3dd6482170ad7a843e3028433154dc40323d8
was scary: it caused us to write keys whose private key did not
correspond to the public key. The key-rotator didn't catch this because
it was trying to treat keys as basically an opaque value that can be
parsed/serialized in various ways, as opposed to a key that can be used
for various cryptographic operations.

Add validations to P-256 values created from any source. Some of the
checks are currently redundant (e.g. the `UnmarshalBinary` path will
currently check if the public key corresponds to a point on the curve
twice), but I leave potential performance enhancements for the future.