Add GCP Secret Manager-based key storage implementation.
Each key version is written to GCP's secret manager as a new version,
effectively using a single secret as an append-only log of our keys.
This is intended for use in a backup solution: keys stored this way
can't be used by the facilitator. This commit is missing one important
piece: we unconditionally try to create the key we will add a version
to; I need to detect "key already exists" as an error, but the error
conditions of this API are not documented AFAICT(!). I'll need to do
some practical testing before I can finish that bit.