Add AWS Secerts Manager-based key storage implementation.
Like the GCP implementation, each new key version will be written to a
new version of a single AWS secret, i.e. we are using an AWS secret as
an append-only log of our secret versions. And like the GCP
implementation, this implementation is suitable for backup, as keys
stored in this way can't be used by the facilitator.