GCP storage.Key: treat already-existing secret as a non-error.
Similar to the AWS case, we blindly try to create keys on write. We want
to treat "key already exists" as a non-error (since in almost all cases
the key will already exist, and we just want to add a new version to the
already-existing key). But I had to do a bit of research to figure out
how to introspect GCP errors properly:
https://pkg.go.dev/cloud.google.com/go#hdr-Inspecting_errors