Key rotator: add flags to force writing keys back to storage. (#1185)
Normally, we only write keys to storage if we detect that they have
actually changed, e.g. the rotation policy caused a new key version to
be created, a new primary version to be elected, etc. Being able to
force a write even when there are no changes would be useful in some
circumstances, e.g. to force a backup, or a re-serialization of existing
keys.