Update handling of serialized private key values. (#1186)
Previously, we serialized P-256 private keys to *up to* 32 bytes. If the
most-significant bytes of the private key happened to be 0, we would
happily elide them and serialize a <32-byte private key value. Relevant
Go playground: https://go.dev/play/p/HRb-cCUgWHb

Note that this change will cause us to hand the facilitator a 32-byte
private key where we would previously have handed the facilitator a
<32-byte private key, due to the chage to AsX962Uncompressed. I think
that's what we want: due to the implementation of decrypt_share and
specifically its use of FixedSliceRandom[1], if the private key bytes
are of a length different than 32, we'll see a panic. (I didn't confirm
with smaller keys, but I have seen this panic from the base64
serialization bug that created 34-byte private keys.)

[1] https://docs.rs/prio/0.5.0/src/prio/encrypt.rs.html#124-127