terraform: configure on demand node group for EKS
This is a pretty hacky fix for the issue encountered in #1192, where
we couldn't allocate spot instances in AWS and found ourselves with 9
nodes in a single AZ, leaving `prometheus-server` unschedulable because
it happened to not be the AZ where its EBS volume is.

To guarantee that we will have at least one node in each AZ, we
configure two node groups: one with cheaper but less reliable spot
instances that can be autoscaled by the `cluster-autoscaler` and another
with what AWS calls "on demand" instances, containing as many nodes as
we have subnets in the cluster (and hence AZs).

Resolves #1192