Facilitator: don't re-parse schemas with every read/write.
Instead, parse schemas once at startup (in a lazy_static block), then
reuse the existing schemas for the remainder of program execution. Other
than being (almost-certainly-negligibly) more performant, this enables
us to parse the schemas using `Schema::parse_list`, which allows the
schemas to include references to one another. The ability to
cross-reference will be important in the next step, which is to
introduce a single message comrpised of the (signature, header, packets)
currently written as three separate messages to three separate cloud
storage objects.