Wire up Terraform to provide default manifest templates.
This implementation strategy provides the key-rotator with the exact
manifests currently included in terraform output, modulo dropping the
(empty) BSK/PEK fields.

The key-rotator does not need the empty BSK/PEK, and I believe these
bogus values might trip some of its validations, so I decided it was a
little cleaner to not have the key-rotator need to deal with them at
all.

Sadly, previous commits didn't consider that the manifest template needs
to be on a per-ingestor basis, so my next step is to go back and teach
the key-rotator to deal with a map of default values.