Key rotator: allow automatic turnup of new localities. (#1220)
This change allows the `key-rotator` to automatically handle provisioning new keys, without the need for an operator to manually run deploy-tool. Once this is deployed to production, we can delete `deploy-tool` entirely.

The main change in this commit is to wire the `key-rotator` to have "manifest templates" so that it knows all the right non-key data to include in a new manifest, and to update the logic to use the templates if no manifest has been written to the appropriate cloud storage bucket.

A small change was also made to the key-rotation logic: if key rotation is disabled, we still need to "rotate" the initial empty key to provision an initial key version for new localities. To support this, we always rotate empty keys, even if key rotation is otherwise disabled for that key type.