Facilitator: add `PrioValidityBatch` message with serialization support (#1216)
This is a step towards making validation batch writes atomic (https://github.com/abetterinternet/prio-server/issues/362).

`PrioValidityBatch` is comprised of a `PrioBatchSignature`, a `PrioValidityHeader`, and a sequence of `PrioValidityPackets` (i.e. the three messages comprising a validation batch). The latter two messages are included as serialized bytes to allow signature verification, and then hash integrity checking, to be done over the same bytes that the signature/hash was generated against.