terraform: module for Kubernetes cluster bootstrap
Introduces a new module in `terraform/cluster_boostrap` which is
responsible for provisioning a managed GKE or EKS Kubernetes cluster
into a GCP project or AWS account provisioned by `isrg-cloud-bootstrap`.

The bulk of the new module consists of the existing `eks` and `gke`
modules, moved to a new location. We do this in a separate TF state to
avoid configuring a TF provider with output from TF managed resources
(see #1046 for details).

The following functionality is new:
 - creation of an AWS ECR or Google Artifact Registry repository for
   container images via Terraform
 - creation of a ClusterRoleBinding for the read-only group in EKS

This commit removes from `terraform/Makefile` management of the GCS
bucket used to store the Terraform remote state. We now expect the
bucket to have been created manually by the operator. This reduces the
risk of bugs in `Makefile` targets accidentally destroying Terraform
state.

Finally, we align documentation in `terraform/README.md` with docs
introduced in isrg-cloud-boostrap [1] and abetterinternet/docs [2].

[1] https://github.com/abetterinternet/isrg-cloud-bootstrap/pull/1
[2] https://github.com/abetterinternet/docs/pull/10

Resolves #1046
Part of #1170
Part of #1031