Specify flags for facilitator-based jobs directly. (#1238)
The preivous configmap-based indirection would not trigger a job restart on configmap change. This has led to a few instances (in preproduction environments, thankfully) where the operator believed the facilitator was running with different config than it was actually running.

Instead, now we specify flags directly in the deployment. This will trigger terraform to cause an update whenever the configuration changes.

Also, remove --pushgateway from facilitator job args. This flag does nothing. I also note it is deprecated in the facilitator code.