Add {batch_header,packet}_bytes to BatchSignature message.
The previous implementation strategy was to introduce a new
ValidationBatch message comprised of a (batch signature, batch header,
batch packets). I realized that this wasn't actually necesary: we can
add new, optional, fields to BatchSignature to achieve the same effect.
The upside of this approach is that reads are significantly simplified:
the old approach would require either attempting to discriminate between
two different message types in a single cloud storage object (a strategy
bound for failure), or more realistically trying to read two separate
objects from cloud storage for each batch (complicating error handling &
requiring more network IO). With this strategy, we read the batch
signature; if it has the header/packet bytes, great, we're done reading;
if not, we read the separate header/packet objects as currently done.

Also remove ValidationBatch since it is no longer necessary.