Facilitator: add --write-single-object-validation-batches flag. (#1261)
If set, the facilitator will emit validation batches in the new
single-object (atomically-written) format. For this to work, the peer
must be able to read single-object batches (i.e. the peer's facilitator
must be built past 457239dd). For that reason, this is off-by-default;
once the relevant change has propagated everywhere, we might consider
turning this on by default.